mirror_updates: introduce fork of S3 mirror sync
This class is supposed to be used at KDE infastructure to allow an
on-the-fly population of sstate mirrors from seeding builds (builds of
the main branches).
In order to avoid a heavy dependency, which the full set of Tegra layers
would be, the 4 files are forked from:

origin: https://github.com/OE4T/tegra-demo-distro.git
hash: 04510ef3128fdeee0946c44eea59d20a027c5540

For documentation and updates see the original locations.